+++ src/main/java/com/landawn/abacus/util/Throwables.java
@@ -2944,4 +2944,44 @@ public final class Throwables {
             void accept(A a, B b, C c) throws E, E2, E3;
         }
     }
+
+    public static final class LazyInitializer<T, E extends Throwable> implements Throwables.Supplier<T, E> {
+        private volatile boolean initialized = false;
+        private volatile T value = null;
+        private final Supplier<T, E> supplier;
+
+        LazyInitializer(final Throwables.Supplier<T, E> supplier) {
+            N.checkArgNotNull(supplier, "supplier");
+
+            this.supplier = supplier;
+        }
+
+        @Override
+        public final T get() throws E {
+            if (initialized == false) {
+                synchronized (this) {
+                    if (initialized == false) {
+                        value = supplier.get();
+
+                        initialized = true;
+                    }
+
+                }
+            }
+
+            return value;
+        }
+
+        /**
+         *
+         * @param <T>
+         * @param supplier
+         * @return
+         */
+        public static <T, E extends Throwable> LazyInitializer<T, E> of(final Throwables.Supplier<T, E> supplier) {
+            N.checkArgNotNull(supplier);
+
+            return new LazyInitializer<>(supplier);
+        }
+    }
 }